# ktoolbox/webui/app.py
from __future__ import annotations

import secrets
from collections.abc import AsyncIterator
from contextlib import asynccontextmanager
from pathlib import Path
from typing import Annotated, cast

from fastapi import Depends, FastAPI, HTTPException, Request, Response, status
from fastapi.exceptions import RequestValidationError
from fastapi.responses import FileResponse, JSONResponse
from fastapi.staticfiles import StaticFiles
from fastmcp.utilities.lifespan import combine_lifespans
from starlette.middleware.base import RequestResponseEndpoint

from ktoolbox import __version__
from ktoolbox.configuration import RuntimeContext
from ktoolbox.project_config import ProjectConfigStore, resolve_project_output
from ktoolbox.webui.about import build_about_response
from ktoolbox.webui.auth import (
    SESSION_COOKIE,
    AuthService,
    client_identifier,
    require_csrf,
    require_session,
)
from ktoolbox.webui.auto_sync_routes import create_auto_sync_router
from ktoolbox.webui.auto_sync_scheduler import AutoSyncScheduler
from ktoolbox.webui.auto_sync_store import AutomaticSyncStore
from ktoolbox.webui.config_monitor import ConfigurationChangeMonitor
from ktoolbox.webui.creator_profiles import CreatorClientFactory, CreatorProfileCache, CreatorRosterService
from ktoolbox.webui.database import WebUIDatabase, WebUISession
from ktoolbox.webui.error_responses import error_payload, validation_error_payload
from ktoolbox.webui.event_store import WebUIEventStore
from ktoolbox.webui.filesystem import FilesystemBrowser
from ktoolbox.webui.filesystem_routes import create_filesystem_router
from ktoolbox.webui.mcp_routes import create_mcp_router
from ktoolbox.webui.mcp_server import create_mcp_server
from ktoolbox.webui.mcp_tokens import MCPTokenStore
from ktoolbox.webui.media import MediaProxyService
from ktoolbox.webui.media_routes import create_media_router
from ktoolbox.webui.models import (
    AboutResponse,
    HealthResponse,
    LoginRequest,
    ProjectSummaryResponse,
    SessionResponse,
)
from ktoolbox.webui.naming_routes import create_naming_router
from ktoolbox.webui.naming_service import NamingConversionService
from ktoolbox.webui.openapi_contract import build_openapi_schema, stable_operation_id
from ktoolbox.webui.pawchive_routes import create_pawchive_router
from ktoolbox.webui.project_lock import ProjectProcessLock
from ktoolbox.webui.project_routes import create_project_router
from ktoolbox.webui.task_executor import TaskExecutor
from ktoolbox.webui.task_routes import create_task_router
from ktoolbox.webui.task_scheduler import TaskScheduler
from ktoolbox.webui.task_store import TaskStore


def create_app(
    context: RuntimeContext,
    *,
    task_executor: TaskExecutor | None = None,
    creator_client_factory: CreatorClientFactory | None = None,
    filesystem_browser: FilesystemBrowser | None = None,
    media_proxy: MediaProxyService | None = None,
) -> FastAPI:
    database = WebUIDatabase(context.project_root / ".ktoolbox" / "webui.sqlite3")
    internal_token = f"mcp-internal-{secrets.token_urlsafe(36)}"
    auth = AuthService(context.configuration.webui, database, internal_token=internal_token)
    event_store = WebUIEventStore(database)
    mcp_token_store = MCPTokenStore(database, event_store)
    task_store = TaskStore(database, event_store)
    automatic_sync_store = AutomaticSyncStore(database, event_store)
    naming_service = NamingConversionService(
        context.project_root,
        database,
        event_store,
        task_store,
    )
    task_scheduler = TaskScheduler(
        context,
        task_store,
        max_concurrency=context.configuration.webui.max_active_tasks,
        executor=task_executor,
    )
    automatic_sync_scheduler = AutoSyncScheduler(
        context,
        task_scheduler,
        task_store,
        automatic_sync_store,
    )
    creator_roster = CreatorRosterService(
        ProjectConfigStore(context.project_root / "ktoolbox.toml"),
        CreatorProfileCache(database, event_store),
        client_factory=creator_client_factory,
    )
    media_service = media_proxy or MediaProxyService()
    project_lock = ProjectProcessLock(context.project_root / ".ktoolbox" / "webui.lock")

    async def update_runtime_context(updated: RuntimeContext) -> None:
        previous = cast(RuntimeContext, app.state.runtime_context)
        app.state.runtime_context = updated
        await naming_service.record_published_time_change(
            previous.configuration.published_time.policy(),
            updated.configuration.published_time.policy(),
        )

    config_monitor = ConfigurationChangeMonitor(
        context.project_root,
        event_store,
        update_runtime_context,
    )

    @asynccontextmanager
    async def lifespan(_: FastAPI) -> AsyncIterator[None]:
        auth.validate_configuration()
        await project_lock.acquire()
        try:
            await database.initialize()
            await naming_service.start()
            await media_service.start()
            await config_monitor.start()
            await task_scheduler.start()
            await automatic_sync_scheduler.start()
            yield
        finally:
            await automatic_sync_scheduler.stop()
            await task_scheduler.stop()
            await config_monitor.stop()
            await media_service.stop()
            await naming_service.stop()
            await project_lock.release()

    app = FastAPI(
        title="KToolBox WebUI API",
        version=__version__.lstrip("v"),
        docs_url=None,
        redoc_url=None,
        lifespan=lifespan,
        generate_unique_id_function=stable_operation_id,
    )
    app.openapi = lambda: build_openapi_schema(app)  # type: ignore[method-assign]
    app.state.runtime_context = context
    app.state.database = database
    app.state.auth = auth
    app.state.mcp_token_store = mcp_token_store
    app.state.event_store = event_store
    app.state.task_store = task_store
    app.state.automatic_sync_store = automatic_sync_store
    app.state.automatic_sync_scheduler = automatic_sync_scheduler
    app.state.naming_service = naming_service
    app.state.media_proxy = media_service
    app.state.task_scheduler = task_scheduler
    app.state.config_monitor = config_monitor
    browser = filesystem_browser or FilesystemBrowser(context.project_root)
    app.state.filesystem_browser = browser
    app.include_router(create_project_router(context.project_root, creator_roster, event_store, config_monitor))
    app.include_router(create_filesystem_router(browser, event_store))
    app.include_router(create_pawchive_router())
    app.include_router(create_media_router())
    app.include_router(create_task_router(context.project_root))
    app.include_router(create_auto_sync_router(context.project_root))
    app.include_router(create_naming_router())
    app.include_router(create_mcp_router())

    static_root = Path(__file__).parent / "static"
    assets_root = static_root / "assets"
    if assets_root.is_dir():
        app.mount("/assets", StaticFiles(directory=assets_root), name="webui-assets")

    @app.get("/mcp", include_in_schema=False)
    async def mcp_frontend() -> FileResponse:
        index = static_root / "index.html"
        if not index.is_file():
            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="not found")
        return FileResponse(index, headers={"Cache-Control": "no-cache"})

    @app.middleware("http")
    async def security_headers(request: Request, call_next: RequestResponseEndpoint) -> Response:
        if request.scope["path"] == "/mcp" and request.method != "GET":
            request.scope["path"] = "/mcp/"
        response = await call_next(request)
        response.headers["X-Content-Type-Options"] = "nosniff"
        response.headers["X-Frame-Options"] = "DENY"
        response.headers["Referrer-Policy"] = "same-origin"
        response.headers["Permissions-Policy"] = "camera=(), microphone=(), geolocation=()"
        response.headers["Content-Security-Policy"] = (
            "default-src 'self'; img-src 'self' data:; style-src 'self' 'unsafe-inline'; "
            "script-src 'self'; connect-src 'self'; frame-ancestors 'none'; base-uri 'self'"
        )
        if request.url.path.startswith("/api/"):
            response.headers["Cache-Control"] = (
                "private, no-store" if request.url.path.startswith("/api/v1/media/") else "no-store"
            )
        return response

    @app.get("/api/v1/health", response_model=HealthResponse)
    async def health() -> HealthResponse:
        return HealthResponse()

    @app.post("/api/v1/session/login", response_model=SessionResponse)
    async def login(payload: LoginRequest, request: Request, response: Response) -> SessionResponse:
        result = await auth.login(payload.username, payload.password, client_identifier(request))
        response.set_cookie(
            SESSION_COOKIE,
            result.token,
            httponly=True,
            secure=request.url.scheme == "https",
            samesite="strict",
            max_age=context.configuration.webui.session_absolute_hours * 3600,
            path="/",
        )
        return SessionResponse(
            username=result.session.username,
            csrf_token=result.session.csrf_token,
            created_at=result.session.created_at,
        )

    @app.get("/api/v1/session", response_model=SessionResponse)
    async def current_session(
        session: Annotated[WebUISession, Depends(require_session)],
    ) -> SessionResponse:
        return SessionResponse(
            username=session.username,
            csrf_token=session.csrf_token,
            created_at=session.created_at,
        )

    @app.post("/api/v1/session/logout", status_code=status.HTTP_204_NO_CONTENT)
    async def logout(
        request: Request,
        _: Annotated[WebUISession, Depends(require_csrf)],
    ) -> Response:
        await auth.logout(request.cookies.get(SESSION_COOKIE))
        logout_response = Response(status_code=status.HTTP_204_NO_CONTENT)
        logout_response.delete_cookie(SESSION_COOKIE, path="/")
        return logout_response

    @app.get("/api/v1/project", response_model=ProjectSummaryResponse)
    async def project(
        request: Request,
        _: Annotated[WebUISession, Depends(require_session)],
    ) -> ProjectSummaryResponse:
        current_context = cast(RuntimeContext, request.app.state.runtime_context).snapshot()
        root: Path = current_context.project_root
        configuration = ProjectConfigStore(root / "ktoolbox.toml").load()
        return ProjectSummaryResponse(
            name=root.name,
            root=root,
            project_config=root / "ktoolbox.toml",
            default_output=configuration.default_output,
            resolved_default_output=resolve_project_output(root, configuration),
            dotenv_files=[root / ".env", root / "prod.env"],
            version=__version__,
            published_target_timezone=current_context.configuration.published_time.target_timezone,
        )

    @app.get("/api/v1/about", response_model=AboutResponse)
    async def get_about(
        _: Annotated[WebUISession, Depends(require_session)],
    ) -> AboutResponse:
        return build_about_response()

    @app.exception_handler(ValueError)
    async def value_error_handler(_: Request, error: ValueError) -> JSONResponse:
        code = status.HTTP_422_UNPROCESSABLE_CONTENT
        return JSONResponse(status_code=code, content=error_payload(str(error), code))

    @app.exception_handler(RequestValidationError)
    async def request_validation_error_handler(_: Request, error: RequestValidationError) -> JSONResponse:
        code = status.HTTP_422_UNPROCESSABLE_CONTENT
        return JSONResponse(status_code=code, content=validation_error_payload(error.errors(), code))

    @app.exception_handler(HTTPException)
    async def http_error_handler(_: Request, error: HTTPException) -> JSONResponse:
        return JSONResponse(
            status_code=error.status_code,
            content=error_payload(error.detail, error.status_code),
            headers=error.headers,
        )

    mcp = create_mcp_server(app, mcp_token_store, internal_token)
    mcp_app = mcp.http_app(path="/", json_response=True, stateless_http=True)
    app.state.mcp = mcp
    app.router.lifespan_context = combine_lifespans(lifespan, mcp_app.lifespan)
    app.mount("/mcp", mcp_app, name="mcp")

    @app.get("/{path:path}", include_in_schema=False)
    async def frontend(path: str) -> FileResponse:
        index = static_root / "index.html"
        if path.startswith("api/") or not index.is_file():
            raise HTTPException(status_code=status.HTTP_404_NOT_FOUND, detail="not found")
        return FileResponse(index, headers={"Cache-Control": "no-cache"})

    return app