# webui/openapi.yaml
openapi: 3.1.0
info:
  title: KToolBox WebUI API
  description: Authenticated REST API for the KToolBox WebUI. Selected operations also define the curated MCP tool surface
    through x-ktoolbox-mcp metadata.
  version: 1.0.0
  license:
    name: BSD-3-Clause
    identifier: BSD-3-Clause
paths:
...
  /api/v1/session/login:
    post:
      summary: Log in
      operationId: login
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      description: Authenticate the configured WebUI account and create a browser session.
      tags:
      - session
      security: []
      x-ktoolbox-mcp:
        enabled: false
        name: login
        scope: null
        safety: null
        openWorld: false
  /api/v1/session:
    get:
      summary: Get current session
      operationId: current_session
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SessionResponse'
      description: Return the authenticated WebUI browser session.
      tags:
      - session
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: current_session
        scope: null
        safety: null
        openWorld: false
  /api/v1/session/logout:
    post:
      summary: Log out
      operationId: logout
      responses:
        '204':
          description: Successful Response
      description: Revoke the current WebUI browser session.
      tags:
      - session
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: logout
        scope: null
        safety: null
        openWorld: false
      parameters:
      - name: X-CSRF-Token
        in: header
        required: true
        description: CSRF token returned by the current browser session.
        schema:
          type: string
          minLength: 1
  /api/v1/project:
    get:
      summary: Get project summary
      operationId: project
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectSummaryResponse'
      description: Return the active project root, configuration files, and KToolBox version.
      tags:
      - project
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: true
        name: get_project_summary
        scope: mcp:read
        safety: read
        openWorld: false
  /api/v1/about:
    get:
      summary: Get application information
      operationId: get_about
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AboutResponse'
      description: Return KToolBox package, license, runtime, author-name, and official-resource information.
      tags:
      - project
      security:
      - WebUISession: []
      x-ktoolbox-mcp:
        enabled: false
        name: get_about
        scope: null
        safety: null
        openWorld: false
components:
  schemas:
    AboutResponse:
      properties:
        name:
          type: string
          title: Name
        version:
          type: string
          title: Version
        description:
          type: string
          title: Description
        license:
          type: string
          title: License
        authors:
          items:
            type: string
          type: array
          title: Authors
        python_version:
          type: string
          title: Python Version
        urls:
          additionalProperties:
            type: string
          type: object
          title: Urls
      type: object
      required:
      - name
      - version
      - description
      - license
      - authors
      - python_version
      - urls
      title: AboutResponse
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    LoginRequest:
      properties:
        username:
          type: string
          maxLength: 256
          minLength: 1
          title: Username
        password:
          type: string
          maxLength: 4096
          minLength: 1
          title: Password
      type: object
      required:
      - username
      - password
      title: LoginRequest
...
    ProjectSummaryResponse:
      properties:
        name:
          type: string
          title: Name
        root:
          type: string
          format: path
          title: Root
        project_config:
          type: string
          format: path
          title: Project Config
        default_output:
          type: string
          format: path
          title: Default Output
        resolved_default_output:
          type: string
          format: path
          title: Resolved Default Output
        dotenv_files:
          items:
            type: string
            format: path
          type: array
          title: Dotenv Files
        version:
          type: string
          title: Version
        published_target_timezone:
          type: string
          title: Published Target Timezone
      type: object
      required:
      - name
      - root
      - project_config
      - default_output
      - resolved_default_output
      - dotenv_files
      - version
      - published_target_timezone
      title: ProjectSummaryResponse
...
    SessionResponse:
      properties:
        authenticated:
          type: boolean
          title: Authenticated
          default: true
        username:
          type: string
          title: Username
        csrf_token:
          type: string
          title: Csrf Token
        created_at:
          type: string
          format: date-time
          title: Created At
      type: object
      required:
      - username
      - csrf_token
      - created_at
      title: SessionResponse
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
...
  securitySchemes:
    WebUISession:
      type: apiKey
      in: cookie
      name: ktoolbox_session
      description: Opaque browser session created by the login operation.
...